CV-2 removes operator and adjustments
removes the | operator. by doing this i had to adjust the cv definition
to be defined as a dictionary list of strings. if i just presented it
as a optional list, it would blow up in the terminal when runnning.
I then changed the jinja to unpack the dictionary by using a for loop
where the outer loop will go through each dictionary such as frontend:
backend: etc.

the inner loop then unpacks the dictionary into a name key and items as
a value. so the name would be frontend. and items are the content inside
of the name. eg: "Frontend: Javascript"